Adding files via upload
I had a problem trying to upload the project as we normally do, so I did it via "upload". The error that was shown while I tried to upload the project, appeared when I tried to make the push, it said "Fatal: the remote end hung up unexpectedly" and I didn't knew how to solve it but I didn't wanted to bother you because I know it is late.